Add files via upload
Implements notification images, reboots USB bus if camera isn't found, adds error checking.

+++ allsky.sh
@@ -6,11 +6,22 @@ then
 fi
 
 # reset auto camera selection, so $ALLSKY_HOME/config.sh do not pick up old camera selection
+cd $ALLSKY_HOME
 echo "" > "$ALLSKY_HOME/autocam.sh"
 source $ALLSKY_HOME/config.sh
 
+echo "     ***** Starting AllSky *****"	# ECC added to make it easy to find beginning of this run in the log file
 echo "Making sure allsky.sh is not already running..."
 ps -ef | grep allsky.sh | grep -v $$ | xargs "sudo kill -9" 2>/dev/null
+
+mv -f log.txt OLD_log.txt 2> /dev/null		# ECC added
+> log.txt
+
+# ECC added notification images
+USE_NOTIFICATION_IMAGES=$(jq -r '.notificationimages' "$CAMERA_SETTINGS")
+if [ "$USE_NOTIFICATION_IMAGES" = "1" ] ; then
+	$ALLSKY_HOME/scripts/copy_notification_image.sh "StartingUp" 2>&1
+fi
 
 # old/regular manual camera selection mode => exit if no requested camera was found
 RPiHQIsPresent=$(vcgencmd get_camera)
@@ -20,11 +31,26 @@ echo "RPiHQ Camera not found. Exiting." >&2
         exit 0
 fi
 
-ZWOIsPresent=$(lsusb -D $(lsusb | awk '/ 03c3:/ { bus=$2; dev=$4; gsub(/[^0-9]/,"",dev); print "/dev/bus/usb/"bus"/"dev;}') | grep -c 'iProduct .*ASI[0-9]')
+# ECC: added: 2>/dev/null, and separated into 2 commands to better aid debugging when camera isn't found.
+ZWOdev=$(lsusb | awk '/ 03c3:/ { bus=$2; dev=$4; gsub(/[^0-9]/,"",dev); print "/dev/bus/usb/"bus"/"dev;}')
+ZWOIsPresent=$(lsusb -D ${ZWOdev} 2>/dev/null | grep -c 'iProduct .*ASI[0-9]')
 if [[ $CAMERA == "ZWO" &&  $ZWOIsPresent -eq 0 ]]; then
-        echo "ZWO Camera not found. Exiting." >&2
-        sudo systemctl stop allsky
-        exit 0
+        echo "ZWO Camera not found..." >&2
+	if [[ $ZWOdev == "" ]]; then
+		echo "  and no USB entry found for it." >&2
+	else
+		echo "  but entry '$ZWOdev' found for it." >&2
+	fi
+	if true; then
+        	echo "  Fixing USB and restarting." >&2
+		sudo ~/uhubctl/uhubctl -a cycle -l 2		# l-ocation 2 are USB 3 ports
+		# No need for 'sudo systemctl stop restart' since the service will automatically restart this script.
+        	exit 0
+	else
+        	echo "  Exiting." >&2
+        	sudo systemctl stop allsky
+        	exit 0
+	fi
 fi
 
 # CAMERA AUTOSELECT
@@ -60,7 +86,6 @@ echo "export CAMERA=$CAMERA" > "$ALLSKY_HOME/autocam.sh"
 source $ALLSKY_HOME/scripts/filename.sh
 
 echo "Starting allsky camera..."
-cd $ALLSKY_HOME
 
 # Building the arguments to pass to the capture binary
 ARGUMENTS=""
@@ -77,10 +102,24 @@ if [[ $1 == "preview" ]] ; then
 fi
 ARGUMENTS="$ARGUMENTS -daytime $DAYTIME"
 
-echo "$ARGUMENTS">>log.txt
+echo "$ARGUMENTS" >> log.txt			# ECC modified
 
+RETCODE=0
 if [[ $CAMERA == "ZWO" ]]; then
 	$ALLSKY_HOME/capture $ARGUMENTS
+	RETCODE=$?
+	echo "capture exited with retcode=$RETCODE"
+
 elif [[ $CAMERA == "RPiHQ" ]]; then
 	$ALLSKY_HOME/capture_RPiHQ $ARGUMENTS
+	RETCODE=$?
+else
+	exit 1
 fi
+
+if [ "$USE_NOTIFICATION_IMAGES" = "1" -a "$RETCODE" -ne 0 ] ; then
+	# "capture" will do this if it exited with 0.
+	$ALLSKY_HOME/scripts/copy_notification_image.sh "NotRunning" 2>&1
+fi
+
+exit $RETCODE
